pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aictrl"
version = "0.1.0"
description = "Build tool for AI coding skills — compiles skill definitions into tool-specific output"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[project.scripts]
aictrl = "aictrl.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aictrl = ["templates/**/*.j2"]
